Added XML exporting. Fixed internationalization.

+++ vogon-data/src/main/java/org/zlogic/vogon/data/FinanceTransaction.java
@@ -199,7 +199,7 @@ public abstract class FinanceTransaction implements Serializable {
 	 *
 	 * @return the transaction amount
 	 */
-	public double getRawAmount() {
+	public long getRawAmount() {
 		return amount;
 	}
